pyproject.toml
[project]
name = "ai-clean-names"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "cdyeq12", email = "xawsolevo@gmail.com" }
]
dependencies = [
    "anthropic>=0.33.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ai_clean_names"]
